Fix verification scoring bugs and add comprehensive tests
Six bugs found via code review and local testing:
- airport_model.airports.get() → get_airport() (wrong API)
- apt.latitude → apt.latitude_deg (wrong attribute)
- ceiling_delta_ft Integer → Float (silent truncation)
- Replace duplicate _haversine_nm with euro_aip NavPoint
- Fix naive/aware datetime subtraction from SQLite
- Fix case-insensitive flight category comparison (VFR vs vfr)

80 new tests covering collection pipeline (27) and scoring (53).

+++ src/weatherbrief/db/models.py
@@ -292,7 +292,7 @@ class VerificationScoreRow(Base):
     category_match: Mapped[bool | None] = mapped_column(Boolean, nullable=True)
 
     # Quantitative deltas (model - observation)
-    ceiling_delta_ft: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    ceiling_delta_ft: Mapped[float | None] = mapped_column(Float, nullable=True)
     visibility_delta_m: Mapped[float | None] = mapped_column(Float, nullable=True)
     wind_speed_delta_kt: Mapped[float | None] = mapped_column(Float, nullable=True)
     wind_dir_delta_deg: Mapped[float | None] = mapped_column(Float, nullable=True)
@@ -346,7 +346,7 @@ class TafVerificationScoreRow(Base):
     category_match: Mapped[bool | None] = mapped_column(Boolean, nullable=True)
 
     # Deltas (TAF - observation)
-    ceiling_delta_ft: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    ceiling_delta_ft: Mapped[float | None] = mapped_column(Float, nullable=True)
     visibility_delta_m: Mapped[float | None] = mapped_column(Float, nullable=True)
     wind_speed_delta_kt: Mapped[float | None] = mapped_column(Float, nullable=True)
     wind_dir_delta_deg: Mapped[float | None] = mapped_column(Float, nullable=True)
